Commit #3
In the NV, on line 22 "Contact Manager" has been modified to "Card Layout". On line 23 the comment "Creating main frame" has been added. On line 27 the comment "Setting the size of the frame" has been added. On line 40 the comment "Adding cardPanel to the frame" has been added. On line 43 the comment "Setting visibility parameter of the frame" has been added.